requirements.txt
pyshark
protobuf
pycryptodome
loguru
